Refactored the DTOS
